wip: add deposit wireformat and utilities
