+++ sbtc-core/Cargo.toml
@@ -7,6 +7,6 @@ edition = "2021"
 log.workspace = true
 once_cell.workspace = true
 regex.workspace = true
-stacks-core.workspace = true
+stacks-core.path = "../stacks-core"
 stacks-rs.workspace = true
 thiserror.workspace = true
