+++ sbtc-core/src/lib.rs
@@ -1,2 +1,1 @@
-pub mod utils;
 pub mod wireformat;
